requirements.txt
flask
requests
bs4
pillow